Move notice components to `Notices`, barrel export; add unit test
