+++ wp-modules/app/js/src/components/Notices/index.ts
@@ -1,0 +1,2 @@
+export { default as EnvironmentNotice } from './EnvironmentNotice';
+export { default as VersionControlNotice } from './VersionControlNotice';
